refactor: merge all functionality into custom ddb client
